testing/selenium/requirements.txt
pytest>=7.0
selenium>=4.10
python-dotenv>=1.0
